gnark-whir: deseralize PK and VK
Creating keys with Setup() is unsafe, because the process is possibly
deterministic which can lead to recovering keys by the attacker. While
it is fine for tetsing purposes, production environments should use keys
derived from a proper multi-party computation trusted setup ceremony.

Add --pk and --vk flags to allow using such keys for verifying the
proof. These flags are optional and not passing any of the keys results
in a fallback to the previous behavior of unsafe keys generation.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>